cleaned up payload, and made the 2 payloads into one folder
