+++ src/main/java/SecondRoll/demo/payload/request/SignupRequest.java
@@ -1,4 +1,4 @@
-package SecondRoll.demo.securitypayload.request;
+package SecondRoll.demo.payload.request;
 
 
 import jakarta.validation.constraints.Email;
