+++ src/main/java/SecondRoll/demo/payload/response/UserInfoResponse.java
@@ -1,38 +1,19 @@
-package SecondRoll.demo.securitypayload.response;
+package SecondRoll.demo.payload.response;
 
 import java.util.List;
 
-public class JwtResponse {
-    private String token;
-    private String type = "Bearer";
+public class UserInfoResponse {
     private String id;
     private String username;
     private String email;
     private List<String> roles;
 
-
-    public JwtResponse(String token, String id, String username, String email, List<String> roles) {
-        this.token = token;
+    public UserInfoResponse(String id, String username, String email, List<String> roles) {
         this.id = id;
         this.username = username;
         this.email = email;
         this.roles = roles;
-    }
 
-    public String getToken() {
-        return token;
-    }
-
-    public void setToken(String token) {
-        this.token = token;
-    }
-
-    public String getType() {
-        return type;
-    }
-
-    public void setType(String type) {
-        this.type = type;
     }
 
     public String getId() {
